added exceptions tests
Issue #3

the exit value is currently failing, though.
Capture::Tiny::capture is storing the exit value as 256 instead of the 1 that's being returned to the shell.
running manually through the shell is storing exit value 1 into $?, however.
will need to research why how to accurately capture and test the exit value here.